+++ packages/toolkit/scripts/check-engines.js
@@ -11,11 +11,7 @@ const { getArgsFromCLI, hasArgInCLI } = require('../utils');
 
 const args = getArgsFromCLI();
 
-const hasConfig =
-	hasArgInCLI('--package') ||
-	hasArgInCLI('--node') ||
-	hasArgInCLI('--npm') ||
-	hasArgInCLI('--yarn');
+const hasConfig = hasArgInCLI('--package') || hasArgInCLI('--node') || hasArgInCLI('--npm') || hasArgInCLI('--yarn');
 const config = !hasConfig ? ['--node', '>=10.0.0', '--npm', '>=6.9.0'] : [];
 
 const result = spawn(resolveBin('check-node-version'), [...config, ...args], {